refactor: `useDevice`를 폴더로 분리
