+++ pesu-components/src/hooks/index.ts
@@ -1,1 +1,1 @@
-export { default as useDevice } from './useDevice';
+export { default as useDevice } from './useDevice/useDevice';
