+++ pesu-components/src/hooks/useDevice/constant.ts
@@ -1,0 +1,6 @@
+export const DEVICES = ['mobile', 'tablet', 'desktop'] as const;
+export const DEVICE_WIDTHS = {
+  mobile: 768,
+  tablet: 1024,
+  desktop: 1024,
+} as const;
