+++ pesu-components/src/hooks/useDevice/type.ts
@@ -1,0 +1,3 @@
+import { DEVICES } from './constant';
+
+export type Device = (typeof DEVICES)[number];
