+++ pesu-components/src/hooks/useDevice/useDevice.ts
@@ -1,16 +1,16 @@
 import { useEffect, useState } from 'react';
 
-const DEVICES = ['mobile', 'tablet', 'desktop'] as const;
-
-export type Device = (typeof DEVICES)[number];
+import { Device } from './type';
+import { DEVICE_WIDTHS } from './constant';
 
 export default function useDevice() {
   const [device, setDevice] = useState<Device>('desktop');
 
   useEffect(() => {
     const handleResize = () => {
-      if (window.innerWidth < 768) setDevice('mobile');
-      else if (window.innerWidth >= 768 && window.innerWidth < 1024) setDevice('tablet');
+      if (window.innerWidth < DEVICE_WIDTHS.mobile) setDevice('mobile');
+      else if (window.innerWidth >= DEVICE_WIDTHS.mobile && window.innerWidth < DEVICE_WIDTHS.tablet)
+        setDevice('tablet');
       else setDevice('desktop');
     };
